Cleaned up the code and organized the file structure
