--- Flask/Templates/install.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<title>Install messageRecording</title>
-<br>
-<br>
-<br>
-<center>
-    <h1>Click the button to install Message Recording Bot</h1>
-    <a href="https://slack.com/oauth/authorize?scope={{ scope }}&client_id={{ client_id }}">
-        <img alt="Add to Slack" height="40" width="139" src="https://platform.slack-edge.com/img/add_to_slack.png" srcset="https://platform.slack-edge.com/img/add_to_slack.png 1x, https://platform.slack-edge.com/img/add_to_slack@2x.png 2x" />
-    </a>
-</center>
