--- Flask/Templates/thanks.html
@@ -1,8 +1,0 @@
-<!doctype html>
-<title>Install messageRecording</title>
-<br>
-<br>
-<br>
-<center>
-    <h1>Thanks for installing!</h1>
-</center>
