+++ SlackMachine/local_settings.py
@@ -1,4 +1,6 @@
-SLACK_API_TOKEN="bot token goes here"
+import os
+
+SLACK_API_TOKEN= os.environ['SLACK_API_TOKEN']
 PLUGINS = [
     'machine.plugins.builtin.general.PingPongPlugin',
     'plugins.listen.EavesdropPlugin'
